LevelManager Test
### Since Last Commit
---
---
#### Additions
- N/A
---
#### Changes
- N/A
---
#### Removed
- N/A

+++ vsSolutions/shadowOperative_001/shadowOperative_001/Enemy.h
@@ -20,7 +20,7 @@ private:
 	bool handleInput();
 public:
 	void update(char** m_ArrayLevel);
-	void spawn(sf::Vector2f startPosition, float gravity);
+	void spawn(sf::Vector2i startPosition, float gravity);
 };
 
 #endif // !ENEMY_H
