+++ vsSolutions/shadowOperative_001/shadowOperative_001/createEnemy.cpp
@@ -1,7 +1,7 @@
 #include "Enemy.h"
 #include "EnemyGenerator.h"
 
-Enemy* createEnemy(sf::Vector2f position,float Gravity)
+Enemy* createEnemy(sf::Vector2i position,float Gravity)
 {
 	Enemy* enemy = new Enemy();
 	enemy->spawn(position, Gravity);
